move model classes from package object to normal classes
