+++ src/main/scala/consul/v1/event/models.scala
@@ -1,0 +1,5 @@
+package consul.v1.event
+
+sealed trait EventIds
+
+case class Event(ID:EventId,Name:String,Payload:Option[String],NodeFilter:String,ServiceFilter:String,TagFilter:String,Version:Double,LTime:Double)
